Move dialog extensions to better namespace
