+++ Source/ClientApplication/RunExcel01.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 using NetOffice;
-using NetOffice.OfficeApi.Tools.Contribution;
 using Excel = NetOffice.ExcelApi;
 using Contribution = NetOffice.OfficeApi.Tools.Contribution;
+using NetOffice.OfficeApi.Extensions.Dialogs;
 
 namespace ClientApplication
 {
