+++ Source/Office.Extensions/Dialogs/AboutDialog.cs
@@ -8,7 +8,7 @@ using System.Text;
 using System.Windows.Forms;
 using NetOffice.OfficeApi.Tools.Contribution;
 
-namespace NetOffice.OfficeApi.Tools.Dialogs
+namespace NetOffice.OfficeApi.Extensions.Dialogs
 {
     /// <summary>
     /// Show about info to the user
