+++ Source/Office.Extensions/Dialogs/AboutDialog.resx
