+++ Source/Office.Extensions/Dialogs/DiagnosticsDialog.resx
