+++ Source/Office.Extensions/Dialogs/PropertyGridDialog.resx
