+++ Source/Office.Extensions/Dialogs/RichTextDialog.resx
