+++ Source/Office.Extensions/Dialogs/ToolsDialog.resx
